qrme: Quantile regression with measurement error

Code for estimating quantile regressions with measurement error. The first main set of functions estimates quantile regression with measurement error in the dependent variable using an EM-type-algorithm. The second set of functions estimates quantile regressions with measurement error in the outcome and in a continuous "treatment" variable. All functions allow for an arbitrary number of other covariates that are assumed not to be measured with error.

See also

Author

Maintainer: Brantly Callaway brantly.callaway@uga.edu

Authors:

Brantly Callaway brantly.callaway@uga.edu

Tong Li tong.li@vanderbilt.edu

Irina Murtazashvili im99@drexel.edu

Emmanuel S. Tsyawo estsyawo@gmail.com