ll

this is the likelihood function for estimating the copula parameter

ll(params, y, t, x, copula = "gumbel", Fyx, Ftx, fyx, ftx, Upi, Umu, Usig,
  Vpi, Vmu, Vsig, ndraws = 100)

Arguments

params
a value of the parameters to calculate the log likelihood for

y
should be a vector of outcomes

t
should be a vector of treatments

x
a matrix of values of x

copula
the type of copula, should be supported by cop.pdf

Fyx
function that contains the cdf of y conditional on x

Ftx
function that contains the cdf of t conditional on x

fyx
function that contains the pdf of y conditional on x

ftx
function that contains the pdf of t conditional on x

Upi
vector of probabilities for measurement error mixture model for Y

Umu
vector of means for measurement error mixture model for Y

Usig
vector of std. deviations for measurement error mixture model for Y

Vpi
vector of probabilities for measurement error mixture model for T

Vmu
vector of means for measurement error mixture model for T

Vsig
vector of std. deviations for measurement error mixture model for T

ndraws
the number of draws of U and V to make

Value

scalar negative value of log likelihood function